requirements.txt
llama-cpp-python
sentence-transformers
pinecone
langchain
langchain-community
langchain-pinecone
flask
python-dotenv
pypdf
-e .